Start treatment for all replay files

+++ LAB_movereplay.py
@@ -58,9 +58,11 @@ if __name__ == '__main__':
                 except:
                     print('Check zipfiles: ',output_dir)
                     continue
-                
-        try:
-            shutil.move(r'{}\morts.csv'.format(input_dir),r'{}\morts.csv'.format(output_dir))
-        except:
-            print('Morts.csv missing for: ',output_dir.split('\\')[-1])
-            
+        
+        if os.path.isfile(r'{}\morts.csv'.format(output_dir)):
+            pass
+        else:
+            try:
+                shutil.move(r'{}\morts.csv'.format(input_dir),r'{}\morts.csv'.format(output_dir))
+            except:
+                print('Morts.csv missing for: ',output_dir.split('\\')[-1])
